update - kakaoAPI, ImageAPI

+++ backend/matching_mate(back)/src/main/java/com/matching/system/service/MatchingPostService.java
@@ -3,13 +3,12 @@ package com.matching.system.service;
 import com.matching.system.domain.*;
 import com.matching.system.dto.ChattingDTO;
 import com.matching.system.dto.MatchingPostDTO;
-import com.matching.system.dto.PagingDTO;
 import com.matching.system.filter.ResponseData;
 import com.matching.system.filter.ResponseMessage;
+import com.matching.system.map.MapControl;
 import com.matching.system.repository.*;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +16,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+//import org.json.JSONArray;
+//import org.json.JSONObject;
 
 @Service
 @Transactional
@@ -30,10 +32,11 @@ public class MatchingPostService {
     private final ChattingService chattingService;
     private final ChattingMemberRepository chattingMemberRepository;
     private final ChattingRoomRepository chattingRoomRepository;
+    @Autowired private final MapControl mapControl;
+    
 
     // 매칭 공고 추가
-    public ResponseMessage save(MatchingPostDTO.CreateDTO matchingPostCreateDTO)
-    {
+    public ResponseMessage save(MatchingPostDTO.CreateDTO matchingPostCreateDTO) {
         // 카테고리 검색
         Optional<Category> category = categoryRepository.findById(matchingPostCreateDTO.getCategoryId());
         if (category.isEmpty()) return new ResponseMessage(HttpStatus.NOT_FOUND, "검색한 카테고리가 없습니다.");
@@ -43,18 +46,20 @@ public class MatchingPostService {
 
         // 저장
         MatchingPost newMatchingPost = matchingPostRepository.save(MatchingPost.builder()
-                                .member(member)
-                                .category(category.get())
-                                .postName(matchingPostCreateDTO.getPostName())
-                                .postContents(matchingPostCreateDTO.getPostContents())
-                                .matchingDate(matchingPostCreateDTO.getMatchingDate())
-                                .matchingTime(matchingPostCreateDTO.getMatchingTime())
-                                .recommendedSkill(matchingPostCreateDTO.getRecommendedSkill())
-                                .numberOfPeople(1)
-                                .maxNumberOfPeople(matchingPostCreateDTO.getMaxNumberOfPeople())
-                                .place(matchingPostCreateDTO.getPlace())
-                                .detailPlace(matchingPostCreateDTO.getDetailPlace())
-                                .build());
+                .member(member)
+                .category(category.get())
+                .postName(matchingPostCreateDTO.getPostName())
+                .postContents(matchingPostCreateDTO.getPostContents())
+                .matchingDate(matchingPostCreateDTO.getMatchingDate())
+                .matchingTime(matchingPostCreateDTO.getMatchingTime())
+                .recommendedSkill(matchingPostCreateDTO.getRecommendedSkill())
+                .isCompleted(0)
+                .numberOfPeople(1)
+                .views(0)
+                .maxNumberOfPeople(matchingPostCreateDTO.getMaxNumberOfPeople())
+                .place(matchingPostCreateDTO.getPlace())
+                .detailPlace(matchingPostCreateDTO.getDetailPlace())
+                .build());
 
         // 채팅 방 추가 + 채팅 회원 추가
         chattingService.createRoom(newMatchingPost, member);
@@ -62,11 +67,11 @@ public class MatchingPostService {
         // 관심 카테고리 member 조회 + notification 저장
         interestCategoryRepository.findByInterestCategoryMember(category.get(), newMatchingPost.getPlace()).stream()
                 .map(interestCategory -> notificationRepository.save(Notification.builder()
-                                .member(interestCategory.getMember())
-                                .notificationType("관심 카테고리 알림")
-                                .message(category.get().getName() + "의 관심 카테고리 공고가 등록되었습니다.")
-                                .url("https://localhost:8080/matchingPost/detail/" + newMatchingPost.getId())
-                                .build()
+                        .member(interestCategory.getMember())
+                        .notificationType("관심 카테고리 알림")
+                        .message(category.get().getName() + "의 관심 카테고리 공고가 등록되었습니다.")
+                        .url("https://localhost:8080/matchingPost/detail/" + newMatchingPost.getId())
+                        .build()
 
                 ))
                 .collect(Collectors.toList());
@@ -75,8 +80,7 @@ public class MatchingPostService {
     }
 
     // 매칭 공고 수정
-    public ResponseMessage update(MatchingPostDTO.UpdateDTO matchingPostUpdateDTO)
-    {
+    public ResponseMessage update(MatchingPostDTO.UpdateDTO matchingPostUpdateDTO) {
         // 카테고리 검색
         Optional<Category> category = categoryRepository.findById(matchingPostUpdateDTO.getCategoryId());
         if (category.isEmpty()) return new ResponseMessage(HttpStatus.NOT_FOUND, "검색한 카테고리가 존재하지 않습니다.");
@@ -85,7 +89,8 @@ public class MatchingPostService {
         Optional<MatchingPost> findMatchingPost = matchingPostRepository.findById(matchingPostUpdateDTO.getId());
         if (findMatchingPost.isEmpty()) return new ResponseMessage(HttpStatus.NOT_FOUND, "검색한 매칭 공고가 존재하지 않습니다.");
 
-        if (findMatchingPost.get().getNumberOfPeople() > matchingPostUpdateDTO.getMaxNumberOfPeople()) return new ResponseMessage(HttpStatus.CONFLICT, "현재 인원보다 더 적은 인원은 입력할 수 없습니다.");
+        if (findMatchingPost.get().getNumberOfPeople() > matchingPostUpdateDTO.getMaxNumberOfPeople())
+            return new ResponseMessage(HttpStatus.CONFLICT, "현재 인원보다 더 적은 인원은 입력할 수 없습니다.");
 
         findMatchingPost.get().updateCategory(category.get());
         findMatchingPost.get().updatePostName(matchingPostUpdateDTO.getPostName());
@@ -113,17 +118,30 @@ public class MatchingPostService {
     }
 
     // 매칭 공고 삭제
-    public ResponseMessage deletePost(Long postId)
-    {
+    public ResponseMessage deletePost(Long postId) {
+        // chatting room : matching post id 삭제
+        Optional<MatchingPost> findMatchingPost = matchingPostRepository.findById(postId);
+        if (findMatchingPost.isEmpty()) return new ResponseMessage(HttpStatus.NOT_FOUND, "검색한 매칭 공고가 존재하지 않습니다.");
+
         matchingPostRepository.deleteById(postId);
 
         return new ResponseMessage(HttpStatus.OK, "정상적으로 삭제했습니다.");
     }
 
     // 매칭 공고 조회 -> 사용자 (최신) -> simple 이냐 모두냐
-    public ResponseData readRecentPosts()
-    {
-        List<MatchingPost> matchingPostList = matchingPostRepository.findByRecentPosts();
+    public ResponseData readRecentPosts(Long categoryId, Double latitude, Double longitude) {
+        String address = mapControl.coordToAddr(longitude, latitude);
+
+        List<MatchingPost> matchingPostList;
+
+        if (categoryId == null) {
+            matchingPostList = matchingPostRepository.findByRecentPosts(address);
+        }
+        else{
+            Category findCategory = categoryRepository.findById(categoryId).get();
+
+            matchingPostList = matchingPostRepository.findByRecentCategoryPosts(findCategory, address);
+        }
 
         List<MatchingPostDTO.ReadDTO> readDTOList = matchingPostList.stream()
                 .map(matchingPost -> changeEntityToDTO(matchingPost))
@@ -133,9 +151,18 @@ public class MatchingPostService {
     }
 
     // 매칭 공고 조회 -> 사용자 (최신) -> simple 이냐 모두냐
-    public ResponseData readPopularPosts()
-    {
-        List<MatchingPost> matchingPostList = matchingPostRepository.findByPopularPosts();
+    public ResponseData readPopularPosts(Long categoryId, Double latitude, Double longitude) {
+        String address = mapControl.coordToAddr(longitude, latitude);
+
+        List<MatchingPost> matchingPostList;
+
+        if (categoryId == null) {
+            matchingPostList = matchingPostRepository.findByPopularPosts(address);
+        } else {
+            Category findCategory = categoryRepository.findById(categoryId).get();
+
+            matchingPostList = matchingPostRepository.findByPopularCategoryPosts(findCategory, address);
+        }
 
         List<MatchingPostDTO.ReadDTO> readDTOList = matchingPostList.stream()
                 .map(matchingPost -> changeEntityToDTO(matchingPost))
@@ -145,22 +172,19 @@ public class MatchingPostService {
     }
 
     // admin 매칭 공고 조회
-    public ResponseData readAdminPosts(PagingDTO paging)
-    {
-        Pageable pageable = PageRequest.of(paging.getFirstPage(), paging.getPageCount());
+    public ResponseData readAdminPosts() {
 
-        List<MatchingPost> matchingPostList = matchingPostRepository.findByRecentPosts();
+        List<MatchingPost> matchingPostList = matchingPostRepository.findAll();
 
         List<MatchingPostDTO.ReadDTO> readDTOList = matchingPostList.stream()
-                    .map(matchingPost -> changeEntityToDTO(matchingPost))
-                    .collect(Collectors.toList());
+                .map(matchingPost -> changeEntityToDTO(matchingPost))
+                .collect(Collectors.toList());
 
         return new ResponseData(HttpStatus.OK, "정상적으로 조회했습니다.", readDTOList);
     }
 
     // 매칭 공고 상세 조회
-    public ResponseData readPostsDetail(Long postId)
-    {
+    public ResponseData readPostsDetail(Long postId) {
         Optional<MatchingPost> findMatchingPost = matchingPostRepository.findById(postId);
 
         if (findMatchingPost.isEmpty()) return new ResponseData(HttpStatus.NOT_FOUND, "검색한 매칭 공고가 존재하지 않습니다.", null);
@@ -168,39 +192,40 @@ public class MatchingPostService {
         // view ++
         findMatchingPost.get().updateViews();
 
-        MatchingPostDTO.ReadDTO readDTO = MatchingPostDTO.ReadDTO.builder()
-                .id(findMatchingPost.get().getId())
-                .memberId(findMatchingPost.get().getMember() == null?null:findMatchingPost.get().getMember().getId())
-                .nickname(findMatchingPost.get().getMember().getNickname())
-                .categoryId(findMatchingPost.get().getCategory().getId())
-                .categoryImgAddress(findMatchingPost.get().getCategory().getImgAddress())
-                .categoryName(findMatchingPost.get().getCategory().getName())
-                .postName(findMatchingPost.get().getPostName())
-                .postContents(findMatchingPost.get().getPostContents())
-                .matchingDate(findMatchingPost.get().getMatchingDate())
-                .matchingTime(findMatchingPost.get().getMatchingTime())
-                .recommendedSkill(findMatchingPost.get().getRecommendedSkill())
-                .numberOfPeople(findMatchingPost.get().getNumberOfPeople())
-                .maxNumberOfPeople(findMatchingPost.get().getMaxNumberOfPeople())
-                .views(findMatchingPost.get().getViews())
-                .place(findMatchingPost.get().getPlace())
-                .detailPlace(findMatchingPost.get().getDetailPlace())
-                .isCompleted(findMatchingPost.get().getIsCompleted())
-                .registerDatetime(findMatchingPost.get().getRegisterDatetime())
-                .build();
+//        MatchingPostDTO.ReadDTO readDTO = MatchingPostDTO.ReadDTO.builder()
+//                .id(findMatchingPost.get().getId())
+//                .memberId(findMatchingPost.get().getMember() == null ? null : findMatchingPost.get().getMember().getId())
+//                .nickname(findMatchingPost.get().getMember().getNickname())
+//                .categoryId(findMatchingPost.get().getCategory().getId())
+//                .categoryImgAddress(findMatchingPost.get().getCategory().getImgAddress())
+//                .categoryName(findMatchingPost.get().getCategory().getName())
+//                .postName(findMatchingPost.get().getPostName())
+//                .postContents(findMatchingPost.get().getPostContents())
+//                .matchingDate(findMatchingPost.get().getMatchingDate())
+//                .matchingTime(findMatchingPost.get().getMatchingTime())
+//                .recommendedSkill(findMatchingPost.get().getRecommendedSkill())
+//                .numberOfPeople(findMatchingPost.get().getNumberOfPeople())
+//                .maxNumberOfPeople(findMatchingPost.get().getMaxNumberOfPeople())
+//                .views(findMatchingPost.get().getViews())
+//                .place(findMatchingPost.get().getPlace())
+//                .detailPlace(findMatchingPost.get().getDetailPlace())
+//                .isCompleted(findMatchingPost.get().getIsCompleted())
+//                .registerDatetime(findMatchingPost.get().getRegisterDatetime())
+//                .build();
+
+        MatchingPostDTO.ReadDTO readDTO = changeEntityToDTO(findMatchingPost.get());
 
         return new ResponseData(HttpStatus.OK, "정상적으로 조회했습니다.", readDTO);
     }
 
-    public MatchingPostDTO.ReadDTO changeEntityToDTO(MatchingPost matchingPost)
-    {
+    public MatchingPostDTO.ReadDTO changeEntityToDTO(MatchingPost matchingPost) {
         return MatchingPostDTO.ReadDTO.builder()
                 .id(matchingPost.getId())
-                .memberId(matchingPost.getMember() == null?null:matchingPost.getMember().getId())
-                .nickname(matchingPost.getMember() == null?null:matchingPost.getMember().getNickname())
-                .categoryId(matchingPost.getCategory() == null?null:matchingPost.getCategory().getId())
-                .categoryName(matchingPost.getCategory() == null?null:matchingPost.getCategory().getName())
-                .categoryImgAddress(matchingPost.getCategory() == null?null:matchingPost.getCategory().getImgAddress())
+                .memberId(matchingPost.getMember() == null ? null : matchingPost.getMember().getId())
+                .nickname(matchingPost.getMember() == null ? null : matchingPost.getMember().getNickname())
+                .categoryId(matchingPost.getCategory() == null ? null : matchingPost.getCategory().getId())
+                .categoryName(matchingPost.getCategory() == null ? null : matchingPost.getCategory().getName())
+                .categoryImgAddress(matchingPost.getCategory() == null ? null : matchingPost.getCategory().getImgAddress())
                 .postName(matchingPost.getPostName())
                 .postContents(matchingPost.getPostContents())
                 .matchingDate(matchingPost.getMatchingDate())
@@ -217,17 +242,16 @@ public class MatchingPostService {
     }
 
     // 채팅방 가입하기
-    public ResponseMessage joinChatting(ChattingDTO.ChattingRoomInOutDTO chattingRoomInOutDTO)
-    {
+    public ResponseMessage joinChatting(ChattingDTO.ChattingRoomInDTO chattingRoomInDTO) {
 
         // 이미 가입 여부 확인
-        Optional<ChattingMember> validateDuplicateRoom = chattingMemberRepository.findByChattingRoomIdAndMemberId(chattingRoomInOutDTO.getRoomId(), chattingRoomInOutDTO.getMemberId());
+        Optional<ChattingMember> validateDuplicateRoom = chattingMemberRepository.findByChattingRoomIdAndMemberId(chattingRoomInDTO.getRoomId(), chattingRoomInDTO.getMemberId());
         if (validateDuplicateRoom.isPresent()) return new ResponseMessage(HttpStatus.CONFLICT, "이미 방에 가입되어 있습니다.");
 
         // chatting_member 추가
-        ChattingRoom chattingRoom = chattingRoomRepository.findById(chattingRoomInOutDTO.getRoomId()).get();
+        ChattingRoom chattingRoom = chattingRoomRepository.findById(chattingRoomInDTO.getRoomId()).get();
 
-        Member member = memberRepository.findById(chattingRoomInOutDTO.getMemberId()).get();
+        Member member = memberRepository.findById(chattingRoomInDTO.getMemberId()).get();
 
         chattingMemberRepository.save(ChattingMember.builder()
                 .chattingRoom(chattingRoom)
@@ -237,6 +261,11 @@ public class MatchingPostService {
 
         return new ResponseMessage(HttpStatus.OK, "정상적으로 처리되었습니다.");
     }
+
+
+
+
+
 
 
 }
